Go: add api mode check in file system command
