+++ cmd/ragflow_cli.go
@@ -69,7 +69,7 @@ func main() {
 			os.Exit(1)
 		}
 	} else {
-		if err = client.NewRun(); err != nil {
+		if err = client.Run(); err != nil {
 			fmt.Printf("CLI error: %v\n", err)
 			os.Exit(1)
 		}
